requirements.txt
fastapi==0.104.1
uvicorn==0.24.0

python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pyjwt==2.8.0
mangum==0.17.0
dnspython==2.4.2
pytest==7.4.3
httpx==0.25.1
pymongo>=4.6.0
motor>=3.3.0

requests == 2.32.3
pytz
apscheduler

# tradestation
# ts-api